miniscript: Move error code to new private error module
There is a whole bunch of error code, move it to a newly created
submodule.

Code move only, no logic changes.
